good

+++ 42/ft_strnstr.c
@@ -6,16 +6,16 @@
 /*   By: mkwizera <mkwizera@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/02 13:56:23 by mkwizera          #+#    #+#             */
-/*   Updated: 2024/03/02 16:52:33 by mkwizera         ###   ########.fr       */
+/*   Updated: 2024/03/13 16:51:38 by mkwizera         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdio.h>
 
-char	*ft_strnstr(const char *str, const char *sub, int n)
+char	*ft_strnstr(const char *str, const char *sub, size_t n)
 {
-	int	i;
-	int	j;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	if (*str == '\0' && n == 0)
